refactor: optimize code stle
