+++ components/form/hooks/useCacheErrors.ts
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import useForceUpdate from '../../_util/hooks/useForceUpdate';
 
 /**
  * Always debounce error to avoid [error -> null -> error] blink
@@ -13,7 +14,7 @@ export default function useCacheErrors(
     visible: !!errors.length,
   });
 
-  const [, forceUpdate] = React.useState({});
+  const forceUpdate = useForceUpdate();
 
   const update = () => {
     const prevVisible = cacheRef.current.visible;
@@ -29,7 +30,7 @@ export default function useCacheErrors(
       prevErrors.length !== errors.length ||
       prevErrors.some((prevErr, index) => prevErr !== errors[index])
     ) {
-      forceUpdate({});
+      forceUpdate();
     }
   };
